requirements.txt
# for cuda 9
# pip install torch==1.2.0+cu92 torchvision==0.4.0+cu92 -f https://download.pytorch.org/whl/torch_stable.html
#
torch
torchvision
bottleneck
tensorflow==1.14
#tensorflow-gpu==1.14
keras==2.2.4
numpy
matplotlib
opencv-python
opencv-contrib-python
scipy
scikit-learn
scikit-image
Pillow
tqdm
pandas
efficientnet
